+++ src/dependency_injector/__init__.py
@@ -1,6 +1,6 @@
 """Top-level package."""
 
-__version__ = "4.46.0"
+__version__ = "4.47.0"
 """Version number.
 
 :type: str